move mod cred helpers to yagu for more general usage
