+++ lib/yagu/netrc.go
@@ -1,4 +1,4 @@
-package git
+package yagu
 
 import (
 	"fmt"
