+++ lib/yagu/netrc_test.go
@@ -1,4 +1,4 @@
-package git
+package yagu
 
 import (
 	"reflect"
